Group create and update loan schemes into 1
